requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
django-model-utils==4.0.0
mysqlclient==2.0.1